requirements.txt
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
numpy==1.26.3
pandas==2.3.1
scipy==1.13.1
tqdm==4.67.1
PyYAML==6.0.2
open3d==0.19.0
matplotlib==3.9.4
plotly==6.3.0
tensorboard==2.20.0
wandb==0.21.1
black==25.1.0
isort==6.0.1
flake8==7.3.0
pyright==1.1.403
python==3.9
shapely
--extra-index-url https://download.pytorch.org/whl/cpu